Changing the value format to Integer
The values in the yAxes and the tooltips are showing as floating point numbers.
They will always be integers so, it will be neat if we format them as integers.

+++ addons/web/static/src/js/views/graph/graph_renderer.js
@@ -261,7 +261,7 @@ return AbstractRenderer.extend({
      */
     _formatValue: function (value) {
         var measureField = this.fields[this.state.measure];
-        var formatter = fieldUtils.format.float;
+        var formatter = fieldUtils.format.integer;
         var formatedValue = formatter(value, measureField, FORMAT_OPTIONS);
         return formatedValue;
     },
